refactor: Liquidation test (#238)
* refactor: Break up liquidation test

* refactor: Liquidation Admin clean up

* feat: Remove smock from liquidation test

* fix: Add missing setup for tests

* chore: Deploy styling

* refactor: Liquidation setup

* refactor: Separate contracts object in Liquidation

* refactor: Destructure contract from helper funcs

* chore: Rename tracerPerps to tracer

* refactor: Break up liquidation receipt tests

* refactor: Liquidate test setup

* refactor: Destructure contract object

* chore: Update test names

* chore: Liquidation test comments

* chore: Rename setup scenarios

* refactor: Name accounts

* refactor: Remove deposit scenario

* chore: Remove unused variables

* refactor: Name accounts in liquidate test

* refactor: Name accounts liquidation admin

* chore: Liquidation test comments
